fix: delete SearchPhotosUi, add setQueryUse, refactoting code
