+++ src/components/SearchPhotosApp/SearchPhotosApp.js
@@ -1,39 +1,114 @@
-import React, { useState } from 'react';
-import SearchPhotosUi from './SearchPhotosUi';
+import React, { useEffect, useState } from 'react';
+import useSearchPhotos from '../useSearchPhotos/useSearchPhotos';
 
-const appAll = {
-  margin: 0,
-  padding: 0,
+const button = {
+  backgroundColor: 'rgba(0, 0, 0, 0.75)',
+  color: 'white',
+  padding: '1rem 2rem',
+  border: '1px solid rgba(0, 0, 0, 0.75)',
+  borderRadius: '20px',
+  fontSize: '1.4rem',
+  cursor: 'pointer',
+  transition: 'background-color 250ms',
 };
 
-const container = {
-  margin: '0 auto',
-  maxWwidth: '1000px',
-  padding: '40px',
+const inputSearch = {
+  fontSize: '1.6rem',
+  padding: '0.5rem 2rem',
+  lineHeight: '2.8rem',
+  borderRadius: '20px',
+  backgroundColor: 'white',
+  marginBottom: '1rem',
 };
 
-const title = {
-  fontSize: '4.4rem',
-  fontFamily: '"Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif',
+const label = {
+  fontSize: '3rem',
+  marginBottom: '1rem',
 };
 
-function SearchPhotosApp() {
-  const [url, setUrl] = useState('');
-  const handleChangeUrl = (value) => {
-    setUrl(value);
+const card = {
+  marginBottom: '1rem',
+  display: 'flex',
+};
+
+const cardImage = {
+  flex: '100%',
+  marginTop: '1rem',
+  borderRadius: '10px',
+};
+
+const cardList = {
+  columnCount: '2',
+};
+
+export default function SearchPhotosApp({ handleChangeUrl }) {
+  const [query, setQuery] = useState('');
+  const [input, setInput] = useState('');
+  const [pics, setPics] = useState([]);
+  const [img, setImg] = useState({});
+
+  const { arrayQuery, queryUse } = useSearchPhotos({
+    key: 'v72jwXBy3QD0RDOkyHAaV2t43JP-Z2IkzA1hBwVf-4M',
+    query: query,
+  });
+
+  useEffect(() => {
+    if (img.id) {
+      handleChangeUrl(img.urls.full);
+    }
+  }, [img.id]);
+
+  useEffect(() => {
+    console.log(arrayQuery);
+    if (arrayQuery.length) {
+      setPics(arrayQuery);
+    }
+  }, [queryUse]);
+
+  const searchPhotos = (e) => {
+    e.preventDefault();
+    setQuery(input);
   };
+
   return (
-    <div className="App" style={appAll}>
-      <div className="container" style={container}>
-        <h1 className="title" style={title}>
-          {'React Photo Search'}
-        </h1>
-        <div className="main">
-          <SearchPhotosUi handleChangeUrl={handleChangeUrl} />
-          <div> URL: {url}</div>
-        </div>
+    <>
+      <form className="form" onSubmit={searchPhotos}>
+        <label className="label" htmlFor="query" style={label}>
+          📷
+        </label>
+        <input
+          type="text"
+          name="query"
+          className="input"
+          style={inputSearch}
+          value={input}
+          onChange={(e) => setInput(e.target.value)}
+        />
+        <button type="submit" className="button" style={button}>
+          Search
+        </button>
+      </form>
+      <div className="card-list" style={cardList}>
+        {pics.map((pic) => (
+          <div
+            className={`card${pic.id === img.id ? ' current-image' : ''}`}
+            style={card}
+            key={pic.id}
+          >
+            <img
+              className="card--image"
+              style={cardImage}
+              alt={pic.alt_description}
+              src={pic.urls.full}
+              width="50%"
+              height="50%"
+              onClick={() => {
+                setImg(pic);
+              }}
+            />
+          </div>
+        ))}
       </div>
-    </div>
+    </>
   );
 }
-export default SearchPhotosApp;
